+++ App/lib/screens/Family/family.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:famealy/blocs/auth_bloc.dart';
 import 'package:famealy/screens/Meal/MealPlan.dart';
+import 'package:famealy/screens/Meal/ShoppingList.dart';
 import 'package:famealy/screens/login/login_screen.dart';
 import 'package:famealy/screens/profile/profile.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -85,7 +86,16 @@ class _FamilyPageState extends State<FamilyPage> {
                           MaterialPageRoute(builder: (context) => mealPlan()),
                         )
                       }),
-              CustomListTile(Icons.list, 'Shopping List', () => {}),
+              CustomListTile(
+                  Icons.list,
+                  'Shopping List',
+                  () => {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) => ShoppingList()),
+                        )
+                      }),
               CustomListTile(Icons.help, 'Tutorial', () => {}),
               CustomListTile(Icons.logout, 'Logout', () => {authBloc.logout()}),
             ],
@@ -119,18 +129,22 @@ class _FamilyPageState extends State<FamilyPage> {
               children: [
                 CustomProfileTile(Icons.group, 'Members', 'FIODJFIJSD'),
                 CustomProfileTile(Icons.account_box, 'Name', 'FIODJFIJSD'),
-                dietryRequirementsTile((){}, Icons.assignment, 'Dietry Requirements'),
+                CustomProfileTile(
+                    Icons.assignment, 'Dietry Requirements', 'FIODJFIJSD'),
                 CustomProfileTile(
                     Icons.announcement, 'Allergies', 'FIODJFIJSD'),
               ],
             )),
-            Container(
-
-            ),
+            Container(),
             Container(
               child: RaisedButton.icon(
                 color: Colors.white,
-                onPressed: () {},
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => mealPlan()),
+                  );
+                },
                 icon: Icon(Icons.fastfood),
                 label: Text('View Weekly Meal Plan'),
               ),
@@ -138,7 +152,12 @@ class _FamilyPageState extends State<FamilyPage> {
             Container(
               child: RaisedButton.icon(
                 color: Colors.white,
-                onPressed: () {},
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => ShoppingList()),
+                  );
+                },
                 icon: Icon(Icons.list),
                 label: Text('View Shopping List'),
               ),
@@ -201,78 +220,5 @@ class _FamilyListWidgetState extends State<FamilyListWidget> {
         },
       ),
     ));
-  }
-}
-
-
-class dietryRequirementsTile extends StatefulWidget {
-  Function onTap;
-  String requirement;
-  IconData icon;
-  String text;
-
-    dietryRequirementsTile(this.onTap, this.icon, this.text);
-
-  @override
-  _dietryRequirementsTile createState() => _dietryRequirementsTile();
-}
-
-class _dietryRequirementsTile extends State<dietryRequirementsTile> {
-  String selectedValue = 'None';
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(8.0, 0, 8.0, 0),
-      child: Container(
-        decoration: BoxDecoration(
-            border: Border(bottom: BorderSide(color: Colors.grey[400]))),
-        child: Container(
-          height: 50,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween, //x axis
-            children: <Widget>[
-              Row(
-                children: <Widget>[
-                  Icon(widget.icon),
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Text(
-                      widget.text,
-                      style: TextStyle(fontSize: 16.0),
-                    ),
-                  ),
-                ],
-              ),
-              DropdownButton<String>(
-                value: selectedValue,
-                icon: Icon(Icons.arrow_downward),
-                iconSize: 24,
-                elevation: 16,
-                style: TextStyle(
-                    color: Colors.deepPurple
-                ),
-                underline: Container(
-                    height: 2,
-                    color: Colors.deepPurple
-                ),
-                onChanged: (String newValue) {
-                  setState(() {
-                    selectedValue = newValue;
-                  });
-                },
-                items: <String>['None', 'Vegan', 'Vegetarian', 'Pescetarian']
-                    .map<DropdownMenuItem<String>>((String value) {
-                  return DropdownMenuItem<String>(
-                      value: value,
-                      child: Text(value)
-                  );
-                }).toList(),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
   }
 }
